Migrate CLI specific code to the binary crate
Alongside the command line version, `envio` also aims to also be a library which provides all of the wonderful
features of the `CLI` version to users.

In order to achieve that all of the the `CLI` specific code has been
migrated from the library crate to the binary crate.

The new API interface is still in its early stages and needs to be
tested out to see if it is simple and easy enough for users to use.