Refactor error handling strategy
Prior to this commit, errors were handled as they occurred. In this
context, "handled" refers to the fact that the error was printed to the
console, and the program exited. This is not ideal, as it makes managing
the codebase much more complicated, and if later down the road we wish
to change the format of these error messages, it would be painstakingly
annoying. This commit changes the error handling to return the error to the
caller. This allows the caller to decide how to handle the error. This
is a more flexible approach and allows for better error handling.