Update target to aarch64-apple-darwin for macOS
Currently there seems to be a issue with the macos-12 runner image when using brew. See [issue 4020](https://github.com/actions/runner-images/issues/4020) for more information.

So for now, support for the x86 architecture is on hold until
the problem is resolved or a alternative is found

macOS user's can now only use envio on the apple silicon chips.