Use EnvVec instead of HashMap to store envs
Instead of storing envs inside of a HashMap in our `Profile` struct,
envio now utilizes a custom type called `EnvVec`, which
encapsulates a Vector of `Env` objects.

Env is another custom type, that stores the data for a specific environment variable such as its name and value.

This change makes it more flexible to include additional fields such as
timestamp, expiration date, comments etc, to a environment variable.

Additionally, instead of only encrypting the environment variables, we serialize the entire `Profile` object and just encrypt that so that when decrypting we do not need to create a new `Profile` object from the decrypted data and just get the entire profile object itself.